feat: added ping cog
